separate out components
